Removing need to fromIntegral internal parsing

+++ src/GameKeeper/API/Queue.hs
@@ -27,8 +27,8 @@ import qualified Data.ByteString.Char8 as BS
 
 data Queue = Queue
     { name      :: BS.ByteString
-    , messages  :: Integer
-    , consumers :: Integer
+    , messages  :: Double
+    , consumers :: Double
     , memory    :: Double
     } deriving (Show)
 
@@ -42,8 +42,8 @@ instance FromJSON Queue where
 
 instance Measurable Queue where
     measure Queue{..} =
-        [ Gauge group (bucket "queue.messages" $ escape name) (fromIntegral messages)
-        , Gauge group (bucket "queue.consumers" $ escape name) (fromIntegral consumers)
+        [ Gauge group (bucket "queue.messages" $ escape name) messages
+        , Gauge group (bucket "queue.consumers" $ escape name) consumers
         , Gauge group (bucket "queue.memory" $ escape name) memory
         ]
